chore: MyRouter와 MemberRouter 병합 #91
