+++ DOTCHI/DOTCHI/Networks/Routers/MemberRouter.swift
@@ -11,6 +11,8 @@ import Moya
 enum MemberRouter {
     case blockUser(userId: Int)
     case reportUser(userId: Int, data: ReportUserRequestDTO)
+    case getMembers(memberId: Int, lastCardId: Int)
+    case patchMembers(data: PatchMembersRequestDTO)
 }
 
 extension MemberRouter: TargetType {
@@ -25,6 +27,10 @@ extension MemberRouter: TargetType {
             return "/blacklists/\(userId)"
         case .reportUser(let userId, _):
             return "/reports/\(userId)"
+        case .getMembers(let memberId, _):
+            return "/members/\(memberId)"
+        case .patchMembers:
+            return "/members/me"
         }
     }
 
@@ -32,6 +38,10 @@ extension MemberRouter: TargetType {
         switch self {
         case .blockUser, .reportUser:
             return .post
+        case .getMembers:
+            return .get
+        case .patchMembers:
+            return .patch
         }
     }
 
@@ -41,6 +51,13 @@ extension MemberRouter: TargetType {
             return .requestPlain
         case .reportUser(_, let data):
             return .requestJSONEncodable(data)
+        case .getMembers(_, let lastCardId):
+            let parameters: [String: Any] = [
+                "lastCardId": lastCardId
+            ]
+            return .requestParameters(parameters: parameters, encoding: URLEncoding.default)
+        case .patchMembers(let data):
+            return .uploadMultipart(data.toMultipartFormData())
         }
     }
