+++ DOTCHI/DOTCHI/Networks/Services/MyService.swift
@@ -11,7 +11,7 @@ import Moya
 class MyViewModel: ObservableObject {
     @Published var myResult: MyResponseDTO?
     
-    private let provider = DotchiMoyaProvider<MyRouter>(isLoggingOn: true)
+    private let provider = DotchiMoyaProvider<MemberRouter>(isLoggingOn: true)
     
     func fetchMy(memberId: Int, lastCardId: Int) {
         provider.request(.getMembers(memberId: memberId, lastCardId: lastCardId)) { result in
@@ -34,7 +34,7 @@ class MyViewModel: ObservableObject {
 class EditViewModel: ObservableObject {
     @Published var editResult: EditResponseDTO?
     
-    private let provider = DotchiMoyaProvider<MyRouter>(isLoggingOn: true)
+    private let provider = DotchiMoyaProvider<MemberRouter>(isLoggingOn: true)
     
     func fetchEdit(data: PatchMembersRequestDTO) {
         provider.request(.patchMembers(data: data)) { result in
